+++ appconfigservice/app/appconfig/config.py
@@ -1,0 +1,4 @@
+APP_CONFIG_MONGO_URL = 'mongodb://localhost:27017'
+APP_CONFIG_DB_NAME = 'app_config_db'
+APP_CONFIG_MAX_POOLSIZE = 100,
+APP_CONFIGS_COLLECTION = 'app_configs'