+++ eventservice/event_rest_service.py
@@ -1,10 +1,12 @@
+import os
+import json
 import logging
 import flask
 
 from bson import ObjectId
 from .db import get_db
 from . import query_params
-from flask import Blueprint, request, make_response, abort
+from flask import Blueprint, request, make_response, abort, current_app
 
 logging.basicConfig(format='%(asctime)-15s %(levelname)-7s [%(threadName)-10s] : %(name)s - %(message)s',
                     level=logging.INFO)
@@ -12,6 +14,18 @@ __logger = logging.getLogger("eventservice")
 
 bp = Blueprint('event_rest_service', __name__, url_prefix='/events')
 
+
+@bp.route('/tags', methods=['GET'])
+def get_tags():
+    response = []
+    try:
+        tags_path = os.path.join(current_app.root_path, "tags.json")
+        with open(tags_path, 'r') as tags_file:
+            response = json.load(tags_file)
+    except Exception as ex:
+        __logger.exception(ex)
+        abort(500)
+    return flask.jsonify(response)
 
 @bp.route('/categories', methods=['GET'])
 def get_categories():