+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add events multiple titles search.
 - Add skip and limit on events query.
 - This CHANGELOG file
 - Events Building Block.
@@ -20,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Profile rest service dockerized
 - Image update and fileDescriptor moved to pii from non-pii
 - Test profiles building block rest service deployed in AWS server
+- Add searching by mobile app version feature to App Config API
+- Add secretKeys to App Config
+- Logging building block API definitions.
+- Auth middleware library.
 
 ### Changed
 - Update events schema to remove endDate as required field and add createdBy and createDate.
@@ -35,8 +40,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Separated non-pii data and pii data and database in profile building block
 - Profile building block rest api changed to Flask RESTFul api service
 - Profile building block non-pii interests has been changed by using category and subcategory
+- Update rokwire.yaml for images entry points of events building block
 - Temporarily ignoring `targetAudience` search query string in events.
+- Changed api doc path from http://.../api/docs to http://.../docs since our domain api.rokwire.illinois.edu contains "api" word.
 
 ### Fixed
 - Events building block API definition to improve text.
 - API documentation in YAML file to reflect the latest code and made text improvements.
+- Issues with API specification YAML file related to app config building block.